Added underwater effects
Added a top layer of water (actually facing down) wich reflects objects
(like the ship and fish) and distorts the above objects (like the skymap
and light rays). Added an animated caustic projection. And added light
rays. I also updated the skymap to be a little brighter/an actual sky.