Updated side tables and Toys
Changed the side tables to be shared between your neighbors. I also
moved in the Control Decks to as close as possible. I also updated the
Toys to be spawned on load with PhotonNetwork.Instantiate so hopefully
they work with multiplayer. Also added in a ground to match the size of
the water overhead. And added in a fog element.