Owner is by PlayArea or grabbing
The owner is now back to whichever play area the object is in or who is
currently holding it. The owner of the object now updates their sync
variables so when transfering ownership the object doesn't go flying
around as much. There is a small bug where users can't pick up other
players toys as the area takes back control right away.