Fixed more bugs for GameOver & Hand Grab if object is destroyed
If the object a hand is grabbing is destroyed the hand will let go of
the object. Game Over should stop all typing of commands and clear the
commands. No score updates should happen when a Game Over is triggered.
Command won't try to be typed if it can't find the object.