Updated ship size & bug fixes
Made the ship smaller, but the ground part a bit bigger. Just so objects
don't go flying around so much. Also updated the lightswitch so that the
animation plays (not sure why but sending the RPC to the Mastermind
caused it to not play the animation as the command to play the animation
was after it.) Also updated the console text to remove any new line
characters coming in so that it can put in its own new line characters.