Added visual global Timer and Console Timer & updated Labels
Added a cylinder that will decrease as level time decreases and change
from green to red. Also added timers above each Console that shrinks as
the command times out; changing from grey to red. These should be
networked. Also updated all Labels from size 45 to size 40 to better fit
text. Updated some of the Labels to fit the double line character limit.
And moved the lightswitch to be centered with the other objects.