Fixed grabbable Grab & spin bug
Updated the Grabbable script to make the GrabMode "Grab" cause items to
keep their position and rotation relative to when they were picked up.
And added Snap to snap the items center to the Hand that picked it up.
Also fixed the spin bug by removing the extra velocity calculation when
the grabbable is dropped since it already has the velocity.