Set Modules back to correct GrabMode & Fixed Pullcord spinning & Grabbed objects
When I added the Snap option it messed up the Modules GrabMode settings.
So I set them back to correct ones. Updated the Pullcord to send its
rotation data as well so it stays up straight. Also added back in the
Grabbable thrown velocity (as just keeping the objects velocity when
it's trying to track towards the hand can cause odd behavior). And
updated the synced Velocity and AngularVelocity to be zero while grabbed
so it updates right where it is grabbed rather than trying to predict.
Meaning there should be less springyness.