PhotonNetwork.Destroy all on scene change
When changing scenes find all objects that have a PhotonView and
PhotonNetwork.Destroy() them (except the OvrRigPhoton, MainCamera, and
Console_Text). This way they are not present for new players. Also
edited the LobbyRoom to instantiate all nessesary objects so that they
are present for new players (because they get Destroyed on loading to
the Game scene). AND enable the TrackedController's before the Hand
scripts so it doesn't throw an error. AND reset level to One when
loading into the LobbyRoom.