GameOver when player leaves during Game
Updated the PhotonGame script to GameOver when a player disconnects.
This is possible by making the Mastermind instantiated for all players
so if the MasterClient switches then the GameOver() function can still
be called. AND the Abort & Reset buttons still work. Updated all module
scripts to correctly get the Mastermind script by Tag (instead of by
name and only for the initial MasterClient). AND updated the
Abort/Reset/Lobby buttons to have GrabMode of "None" (they were never
updated when Snap was introduced).