Update WaitForPlayersToSpawn & Abort/Reset button PhotonView
Updated the WaitForPlayersToSpawn to only wait for the players when
loading into the Game scene (as it should be run right away by the
MasterClient in the LobbyRoom scene.). Also removed the Transform from
the PhotonView's observed componenet of the Abort and Reset buttons as
it doesn't need to be synced.