Vive bug fixes
Fixed the following bugs for Vive. Audio not playing (removed the
PhotonView on the Ears as it was not needed and cuasing them to be
Destroyed). Updated recenter to the Menu button (instead of Trackpad).
Moved hand models down so they better match the players hands. Increased
haptics to last 1 second (instead of 1/2 a second).