Increase LevelTimeoutSeconds & ScoreToLose.
At level ONE LevelTimeoutSeconds is 100 seconds (10 seconds per command), by level 5 this is 85.195 seconds, by level 10 this is 76.493 seconds (converging to 70 seconds by level 25). This was increased due to Hazards being implemented players need a little more time to handle them. Also increased the ScoreToLose to -2 as failing right when the round starts is kind of intimidating if you miss the first one.