MainMenu text updated to show brief errors
The LobbyText will now show an error message for 3 seconds if a user
tries to join a game but fails (no longer exits, full, or invisible).
Updated when the OVRCameraRig gets de-coupled to be when the user
actually joins the game, not when they press the button. In case they
are unable to join the game we can't have their OVRCameraRig floating
around.